drf-auth-and-validation/api_with_restrictions/requirements.txt
Django>=6.0.0,<6.1.0
djangorestframework>=3.16.0
django-filter==25.2
psycopg2-binary>=2.9.3
